updated form

+++ src/components/sections/DoctorList/DoctorItem/DoctorItem.tsx
@@ -1,27 +1,38 @@
 'use client';
 import { FC, SetStateAction, Dispatch } from 'react';
+import styles from './DoctorItem.module.scss';
 import { DoctorType } from '@/types';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
+import { useTranslations } from 'next-intl';
 
 interface Props {
   doctor: DoctorType;
   setCurrentDoctor: Dispatch<SetStateAction<DoctorType>>;
   setIsOpen: Dispatch<SetStateAction<boolean>>;
+  loggedIn: boolean;
 }
 
-const DoctorItem: FC<Props> = ({ doctor, setCurrentDoctor, setIsOpen }) => {
+const DoctorItem: FC<Props> = ({
+  doctor,
+  setCurrentDoctor,
+  setIsOpen,
+  loggedIn,
+}) => {
   const router = useRouter();
+  const t = useTranslations();
   const handleDelete = async () => {
-    await fetch('/api/doctors', {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        _id: doctor._id,
-      }),
-    });
+    const confirmed = window.confirm(`Are you sure to delete ${doctor.name}?`);
+    if (confirmed)
+      await fetch('/api/doctors', {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          _id: doctor._id,
+        }),
+      });
     // const data = await res.json();
     router.refresh();
   };
@@ -32,26 +43,47 @@ const DoctorItem: FC<Props> = ({ doctor, setCurrentDoctor, setIsOpen }) => {
   };
 
   return (
-    <div>
-      <span>{`name 
-    ${doctor.name} 
-    city 
-    ${doctor.city} 
-    contact 
-    ${doctor.contact} 
-    instagram
-    ${doctor.instagram}`}</span>
-      <button onClick={handleEdit}>
-        <Image src="/images/edit.svg" width={20} height={20} alt="edit icon" />
-      </button>
-      <button onClick={handleDelete}>
-        <Image
-          src="/images/delete.svg"
-          width={20}
-          height={20}
-          alt="delete icon"
-        />
-      </button>
+    <div className={styles.container}>
+      <span>{`${t('doctorList.name')}: ${doctor.name}`}</span>
+      {doctor.city.length ? (
+        <span>{`${t('doctorList.city')}: ${doctor.city}`}</span>
+      ) : null}
+      {doctor.profession.length ? (
+        <span>{`${t('doctorList.profession.profession')}: ${
+          doctor.profession === '1'
+            ? t('doctorList.profession.1')
+            : t('doctorList.profession.2')
+        }`}</span>
+      ) : null}
+      {doctor.area.length ? (
+        <span>{`${t('doctorList.area')}: ${doctor.area}`}</span>
+      ) : null}
+      <span>
+        {`${t('doctorList.contact')}: `}
+        <a className={styles.phoneLink} href={`tel:${doctor.contact}`}>
+          {doctor.contact}
+        </a>
+      </span>
+      {loggedIn && (
+        <div className={styles.buttonBox}>
+          <button className={styles.editButton} onClick={handleEdit}>
+            <Image
+              src="/images/edit.svg"
+              width={20}
+              height={20}
+              alt="edit icon"
+            />
+          </button>
+          <button className={styles.deleteButton} onClick={handleDelete}>
+            <Image
+              src="/images/delete.svg"
+              width={20}
+              height={20}
+              alt="delete icon"
+            />
+          </button>
+        </div>
+      )}
     </div>
   );
 };
